refactor: Rework dep. injection for each class
Each class is now responsible for knowing about and instantiating
the class that is their direct and immediate collaborator.
Example: A Game instance needs 2 Player instances so it may create
those 2 Player instances. It does not need to know that Player requires
a Gameboard instance. That knowledge lies in the Player class.
If a class can operate with less knowledge about the outside world, then
that knowledge is extraneous.